MAPpredictEndTime converted to ulong

+++ speeduino/sensors.h
@@ -47,7 +47,7 @@ unsigned long MAPrunningValue; //Used for tracking either the total of all MAP r
 unsigned long EMAPrunningValue; //As above but for EMAP
 unsigned int MAPcount; //Number of samples taken in the current MAP cycle
 uint32_t MAPcurRev; //Tracks which revolution we're sampling on
-uint16_t MAPpredictEndTime; //MAP prediction functionality end time (uses ms_counter)
+unsigned long MAPpredictEndTime; //MAP prediction functionality end time (uses ms_counter)
 bool auxIsEnabled;
 byte TPSlast; /**< The previous TPS reading */
 unsigned long TPS_time; //The time the TPS sample was taken
